feat: DELETE /v1/articles/:slug

+++ src/infrastructure/database/mysql/ArticleRepository.ts
@@ -484,13 +484,11 @@ export default class ArticleRepository implements ArticleRepositoryInterface.def
     });
   }
 
-  public deleteArticle(slug: Slug): Promise<void> {
+  public deleteArticle(slug: Slug, email: Email): Promise<void> {
     return new Promise<void>((resolve, reject) => {
       this._connection.query(
-        'DELETE FROM articles WHERE slug = ?',
-        [
-          slug.value
-        ],
+        'DELETE FROM articles WHERE slug = ? AND author_email = ?',
+        [slug.value, email.string()],
         (err: any | null, result: any) => {
           if (err) {
             console.error(err)
